tests: inherit animation tests
